Add FPS counter via Emigui -- an immediate mode gui
Emigui seemed to be the one most easily implementable without any
existing instance for Luminance, and it kind of was. A couple of
tricky details, but nothing major. All in all quite simple.

Emigui is not exactly feature-complete though. Even something as
simple as monospace fonts is missing at the moment. Maybe I should
help out?